update login module

+++ src/FindMyTro/src/main/java/com/FindMyRoom/repository/AdminRepository.java
@@ -5,8 +5,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AdminRepository extends CrudRepository<Admin, Integer> {
+public interface AdminRepository extends CrudRepository<Admin, Long> {
     List<String> getEmails();
+
     Admin getByEmail(String email);
+
     Admin getById(long id);
 }
