+++ src/FindMyTro/src/main/java/com/FindMyRoom/repository/BusinessRepository.java
@@ -3,5 +3,5 @@ package com.FindMyRoom.repository;
 import com.FindMyRoom.model.Business;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BusinessRepository extends CrudRepository<Business, Integer> {
+public interface BusinessRepository extends CrudRepository<Business, Long> {
 }
