+++ src/FindMyTro/src/main/java/com/FindMyRoom/repository/EmployeeRepository.java
@@ -1,8 +1,7 @@
 package com.FindMyRoom.repository;
 
-import com.FindMyRoom.model.Admin;
 import com.FindMyRoom.model.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
+public interface EmployeeRepository extends CrudRepository<Employee, Long> {
 }
